code-health-audit: first wave of high-impact fixes
Acts on the multi-agent code-health audit run across Data/, Profiling/,
Persistence+Insights, Web/, and UI/. This wave is the critical-priority
slice — correctness bugs, invariant violations, hot-path allocations,
and Invariant-3 wording fixes. Lower-priority NIT-class findings and
larger refactors are documented for a follow-up.

Correctness / invariant fixes:
- SegmentDetector.ComputeBiomeComposite: memoised by bitset; was
  allocating a StringBuilder + final string every tick (Invariant 2
  violation — the biome composite is stable across thousands of
  consecutive ticks).
- BoolIndex.EnsureCapacity: infinite loop when constructed with
  capacity=0 (`0 *= 2` never grows). Floor doubling base at 1.
- DataRegistry.Register/DisposeAll: take the gate lock around BOTH
  the name-map and ordered-list mutations so concurrent register/
  dispose can't observe one without the other.
- AllocationCollector: honour the doc contract — null out the byte
  list refs when the active backend does not track allocations,
  rather than relying on consumers to read the tracks flag.
- SegmentAggregator: tighten the nullable flow (`sys == null` short
  circuit) so the compiler analysis matches the runtime semantics.
- SelfHealthStat: defensive null path on the static fallback.
- PlayerDeathDetector: drop the (short) cast on the boss-id passed
  to Lang.GetNPCNameValue. Truncated modded boss type ids ≥ 32768
  (real on heavy Calamity/Thorium-class modlists).
- ContextTransitionWatcher: encode the weather flag identity in
  the transition Type field ("weather:BloodMoon" not just
  "weather"); pre-fix, every flag flip collapsed into an
  indistinguishable row.
- TickDownsampler.RollingFrame: recompute _max on eviction so the
  window's stated Max actually reflects the rolling window (was
  monotonic-since-session-start, never decreased).
- ModlistStream: derive SessionCount from the Sessions collection
  rather than blindly incrementing, and dedupe VersionHistory
  appends — both fixes make the stream idempotent under journal
  replay.
- InsightsEngine.GetOrCreateShared: CompareExchange the lazy init
  so concurrent first-callers can't allocate twice.
- PerformanceProfiler.cs: delete the orphan duplicate <summary>
  block dangling before RegisterDataPipeline.

Honesty / cadence labelling:
- Data/Collectors/* — switch to OnDemand cadence + null
  PerTickCallback. They were declaring PerTick + a no-op delegate;
  the actual per-tick capture is owned by MetricCollector. OnDemand
  matches who-does-work.

Invariant 3 (descriptive, not normative) wording fixes in dashboard:
- Js.Insights: "FreeRemovalCandidate" label "possibly removable"
  → "idle most of session". "Removable" implied a player decision.
- Js.Lag: empty state "no lag events yet — clean session" →
  "no spikes or stalls observed in the last 30s". "Clean" was
  evaluative.
- Js.ModCard: "if this mod were removed, your frame would be ..."
  → "modelled cost without this mod's contribution ..." with an
  explicit "descriptive of measured cost, not a recommendation"
  caveat. The original framing read as a recommendation.

Performance / allocation:
- DashboardAssets.Css/Js: switched from get-only property
  `=> string.Concat(...)` to `static readonly` initialised once.
- DashboardRouter: cache the UTF-8-encoded CSS / JS bytes once
  at type-init so /dashboard.css and /dashboard.js don't re-encode
  per request (cold-tab refresh was paying tens of KB per hit).
- KpiCalculator: ThreadStatic scratch buffer for the median sort.
  /api/now polls at ~1.5 Hz and was allocating a fresh double[1800]
  on every call.

+++ Profiling/Persistence/PlayerDeathDetector.cs
@@ -61,7 +61,11 @@ internal sealed class PlayerDeathDetector
         for (int i = 0; i < ctx.Bosses.Count; i++) bosses.Add(ctx.Bosses[i]);
 
         string primaryBoss = bosses.Count > 0
-            ? (Lang.GetNPCNameValue((short)bosses[0]) ?? ("npc-" + bosses[0]))
+            // tML's int overload of GetNPCNameValue is the one that handles
+            // modded NPC types beyond short.MaxValue. Casting to short would
+            // truncate any modded boss type id ≥ 32768 — a real possibility
+            // on large modlists like Calamity + Thorium + 50-odd content mods.
+            ? (Lang.GetNPCNameValue(bosses[0]) ?? ("npc-" + bosses[0]))
             : "(no boss)";
 
         string primaryBiome = BiomeRegistry.NameOrIndex(ctx.Biomes.PrimaryBitIndex());
